added confirmation notice & database file

+++ src/com/epustaka/frontend/Anggota.java
@@ -10,7 +10,10 @@ import com.epustaka.backend.Validation;
 import com.formdev.flatlaf.FlatLightLaf;
 import java.sql.Connection;
 import javax.swing.JOptionPane;
+import javax.swing.RowFilter;
 import javax.swing.table.DefaultTableModel;
+import javax.swing.table.TableModel;
+import javax.swing.table.TableRowSorter;
 
 /**
  *
@@ -99,6 +102,9 @@ public class Anggota extends javax.swing.JFrame {
         txtAlamat = new javax.swing.JTextField();
         jLabel6 = new javax.swing.JLabel();
         lbId = new javax.swing.JLabel();
+        jPanel4 = new javax.swing.JPanel();
+        jLabel10 = new javax.swing.JLabel();
+        txtSearch = new javax.swing.JTextField();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("ePustaka | Menu Anggota");
@@ -224,7 +230,7 @@ public class Anggota extends javax.swing.JFrame {
         });
         jScrollPane1.setViewportView(tbAnggota);
 
-        getContentPane().add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 70, 720, 640));
+        getContentPane().add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 140, 720, 570));
 
         jPanel2.setBackground(new java.awt.Color(255, 255, 255));
         jPanel2.setBorder(new javax.swing.border.LineBorder(new java.awt.Color(194, 194, 194), 1, true));
@@ -383,6 +389,47 @@ public class Anggota extends javax.swing.JFrame {
 
         getContentPane().add(jPanel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(740, 70, 530, 640));
 
+        jPanel4.setBackground(new java.awt.Color(255, 255, 255));
+        jPanel4.setBorder(new javax.swing.border.LineBorder(new java.awt.Color(194, 194, 194), 1, true));
+
+        jLabel10.setText("Cari Anggota");
+
+        txtSearch.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtSearchActionPerformed(evt);
+            }
+        });
+        txtSearch.addKeyListener(new java.awt.event.KeyAdapter() {
+            public void keyTyped(java.awt.event.KeyEvent evt) {
+                txtSearchKeyTyped(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
+        jPanel4.setLayout(jPanel4Layout);
+        jPanel4Layout.setHorizontalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(txtSearch)
+                    .addGroup(jPanel4Layout.createSequentialGroup()
+                        .addComponent(jLabel10, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 636, Short.MAX_VALUE)))
+                .addContainerGap())
+        );
+        jPanel4Layout.setVerticalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel10)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(txtSearch, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(8, Short.MAX_VALUE))
+        );
+
+        getContentPane().add(jPanel4, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 70, 720, 60));
+
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
@@ -457,6 +504,11 @@ public class Anggota extends javax.swing.JFrame {
         if (!Validation.validateEmail(txtEmail.getText()) || !Validation.validatePhone(txtTelepon.getText())) {
             return;
         } 
+        
+        int konfirmasi = JOptionPane.showConfirmDialog(null, "Apakah anda yakin mengubah data anggota " + txtNama.getText() + "?", "Pemberitahuan",  JOptionPane.YES_NO_OPTION);
+        if(konfirmasi == 1 || konfirmasi == JOptionPane.CLOSED_OPTION) {
+            return;
+        }
                 
         try {
             String sql = "UPDATE anggota SET nama = '"
@@ -478,6 +530,11 @@ public class Anggota extends javax.swing.JFrame {
 
     private void btHapusActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btHapusActionPerformed
         // TODO add your handling code here:
+        int konfirmasi = JOptionPane.showConfirmDialog(null, "Apakah anda yakin menghapus anggota " + txtNama.getText() + "?", "Pemberitahuan",  JOptionPane.YES_NO_OPTION);
+        if(konfirmasi == 1 || konfirmasi == JOptionPane.CLOSED_OPTION) {
+            return;
+        }
+        
         try {
             String sql = "Delete FROM anggota WHERE id ='" + lbId.getText() + "'";
             java.sql.Connection conn = (Connection)Config.configDB();
@@ -514,6 +571,21 @@ public class Anggota extends javax.swing.JFrame {
         dispose();
     }//GEN-LAST:event_btMenuBukuActionPerformed
 
+    private void txtSearchActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtSearchActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtSearchActionPerformed
+
+    private void txtSearchKeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_txtSearchKeyTyped
+        // TODO add your handling code here:
+        String searchText = txtSearch.getText().trim();
+
+        TableRowSorter<TableModel> rowSorter = new TableRowSorter<>(tbAnggota.getModel());
+        tbAnggota.setRowSorter(rowSorter);
+
+        RowFilter<TableModel, Object> rowFilter = RowFilter.regexFilter("(?i)" + searchText);
+        rowSorter.setRowFilter(rowFilter);
+    }//GEN-LAST:event_txtSearchKeyTyped
+
     /**
      * @param args the command line arguments
      */
@@ -541,6 +613,7 @@ public class Anggota extends javax.swing.JFrame {
     private javax.swing.JButton btMenuPustakawan;
     private javax.swing.JButton btTambah;
     private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel10;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
     private javax.swing.JLabel jLabel4;
@@ -549,12 +622,14 @@ public class Anggota extends javax.swing.JFrame {
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel2;
     private javax.swing.JPanel jPanel3;
+    private javax.swing.JPanel jPanel4;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JLabel lbId;
     private javax.swing.JTable tbAnggota;
     private javax.swing.JTextField txtAlamat;
     private javax.swing.JTextField txtEmail;
     private javax.swing.JTextField txtNama;
+    private javax.swing.JTextField txtSearch;
     private javax.swing.JTextField txtTelepon;
     // End of variables declaration//GEN-END:variables
 }
